chore: make local dev server more resilient (#156)
As shown in #154 when a lambda crashes, the whole express server stops
working. This is not only a pain for local dev, but one could argue that
doesn't even replicate the AWS environment (one cannot crash an API
gateway, and lambda runs are isolated from each other, etc).

I simplified the issue as much as I could, removing the `eslint` stage
(assuming the files had been pre-generated) and realised that `nodemon`
doesn't handle restarts properly, even when it's just the express
server.

So I started using
[pm2](https://pm2.keymetrics.io/docs/usage/quick-start/).

In `pm2` they ask you to define an _ecosystem_ file where you setup the
apps that pm2 will run. I've defined 2 "apps":

1. The `express-local-server` app takes advantage of pm2 ability to
watch and restart on files changes (same as `nodemon` did) but also
takes advantage of the fact that pm2 will keep the server running (a
difference of `nodemon`).

2. Then the `esbuild` app is run regularly (as far as pm2 is aware), but
we're passing a `--watch` flag to it via the CLI. I did it this way
because `esbuild` is already aware of the files that it has to watch:
the same files than it has to bundle. But I'm open to tweaking it.

In the end, the setup works as follows:

1. A `src/` file is modified.
2. `esbuild` (on watch mode) detects that a file in `src/` has been
modified. This triggers a rebuild into `dist/`.
3. When `esbuild` did the rebuild, it modified files in `dist/lambdas/`
which triggers a `express-local-server` restart (through `pm2` this
time).