feat: refresh endpoint  (#151)
Architecture:
- user previously logged in using POST /api/v1/login and was given an
accessToken and a refreshToken. This refresh token was stored in a
DynamoDB table named RefreshTokens-$env indexed by UserId(partitionKey)
and RefreshTokenId(sort key) - i.e. a random uuid that identifies the
refresh token. Reason for this design is that a user might use multiple
devices. Aside from these fields, we store the token itself and the
expiration timestamp in ms in the ExpiredAt field. This is used by
Dynamo Streams(TTL) for housekeeping.
- user sends a request to POST /api/v1/refresh with { refresh-token:
$refreshToken } where $refreshToken is another Jwt with no extra claims
and much longer expiration. The endpoint is unprotected _Obvs_.
 - Endpoint business logic:
1) it attempts to decode it and verify the signature. if success 2)
otherwise 401.
2) it attempts to get the token from storage by UserId and
RefreshTokenId. if found 3) otherwise 500. The reason for this is that
as we have verified the token at the previous step, it is not expected
to get this far if the refresh token has expired.
3) if storedToken == requestProvidedToken, then 4) otherwise 403. We
assume here something malicious has been attempted.
4) it attempts to build fresh accessToken and refreshTokens and store
refreshToken in storage. if success 200OK otherwise 500, cause the
reasons could be as varied as the endpoint config is wrong or dynamoDb
does not let us store the record.
   
  More acomplisments:
 - tf code and openapi spec update to project these changes.
 - now login endpoint returns a valid refreshToken
   
  Bonus points:
- removed some unit test boilerplate that tidies up the end before
running the next unit test in favour of some Jest global setup that
makes sure it happens everytime.
- some error handling in $Service$Stores. More to do on this front as we
address https://github.com/Notifycal/backend/issues/118
   - Added Content-Type: application/json header to everything.
- more work on types and validation, that in turn, enables better types.

---------

Co-authored-by: github-actions <github-actions@github.com>
Co-authored-by: dsiguero <26867647+dsiguero@users.noreply.github.com>