feat: handle senderDetails so it can be both SMS or RCS (#394)
I've added `senderDetails` to the 2nd lambda for now, but I reckon they
should come from the first lambda. Should we use this PR to pass
`senderDetails` from the first lambda? Probably doesn't make sense to
store it in DB yet, but we could hardcode it in the first lambda and
pass it from there?

Also, would it make sense to add tests here to cover `senderDetails`
being SMS or RCS, or that'd be more for the 3rd lambda? Asking as in the
first lambda it's hardcoded, and the 2nd one just does a passthrough