feat: advanced openapi spec bundling with redocly (#418)
It all started because I realised the linter for the OpenAPI spec was
pointing to the wrong file... xD

This unlocks the ability to use multiple files to define the API, and
`redocly` (previously called `swagger-cli`) will merge them. I started
investigating into this because:
- I wanted the messages webhook to live somewhat isolated from the rest
of the API spec.
- I didn't want to have to copy somebody else (Vonage) OpenAPI spec
(requestBody schemas, etc...).

I did a diff between the previously generated file (with `yq`, in
`build.sh`) and the one `redocly` generates and it changes the ordering
a bit, removes comments, and triple single quotes some stuff, but I
couldn't see anything wrong.