fix: add `function_name` default dimension to Powertools metrics (#444)
Looks like the `function_name` dimension in Powertools metrics isn't
included by default. We've seen it filled for the `ColdStart` metric
because Powertools metrics probably internally add the dimension to that
specific metric only. But wasn't displayed by default when doing
`metrics.addMetric('Name', MetricUnit.Count, 1)`

This PR adds the `function_name` metric to the Powertools metric
initialization, so all the metrics will have it as a dimension.

Before:
![Screenshot 2025-03-19 at 10 51
20](https://github.com/user-attachments/assets/f47359d1-9609-4b7b-bc1d-4c0955cc8587)


After: 
![Screenshot 2025-03-19 at 10 52
00](https://github.com/user-attachments/assets/85b82f99-e33c-49e7-a877-e238705f6f2e)