feat: improve logging by adding keys (#510)
correlationId had different meanings depending on the context:
- if lambda invoked by APIGateway, then correlationId ===
requestContext.requestId. This is AWS correlationId.
- if lambda invoked by EventBridge, then correlationId === .id. This is
AWS correlationId.
- if lambda invoked by SQS, then correlationId === Notifycal event
.correlationId.
that's what [Powertools
suggest](https://docs.powertools.aws.dev/lambda/typescript/latest/core/logger/#setting-a-correlation-id:~:text=Description-,API_GATEWAY_REST,-%27requestContext.requestId%27)
and it is kind of in line with how we set correlationId. If action gets
triggered by AWS, we embrace the id and make it ours.

Disclaimer: the latter does not stand for APIGateway. Neither it does
for SQS but SQS processing is always an intermediate or final part of
the processing chain.