feat: email service (#583)
Mind, email content is capped by the 256KiB message size limit from
SNS/SQS. Same thing setup up the non functional requirement for batch
size = 1.

Bonus points:
- abstracted code away for idempotency. Mainly, it has consisted of
extracting the idempotency mechanism to be reused as it is a kind of
fragile piece of code - especially for the time being as we are relying
on manual integration/acceptance testing. Class is named
AbstractIdempotencyProcessor. Then, each idempotent processing you wanna
perform you extend this class to be able to call processIdempotently -
this method takes all the config and required handlers for it to work.
In turn, the IdempotentProcessor instantiates a Processor - which
implements the core logic on how to process a message. Lastly, lamdas
just have to instantiate IdempotentProcessor and invoke
"<sendEmail|sendReminderIdempotently>Idempotently". If explanation got
you head dizzy, have a look at this [UML? I don't remember UML anymore]
diagram.

![image](https://github.com/user-attachments/assets/e70c6b66-62f3-4390-9216-22cb1603aec2)

 - messagingEnabled tf var so we can disable it per env.

---------

Co-authored-by: dsiguero <26867647+dsiguero@users.noreply.github.com>
Co-authored-by: github-actions <github-actions@github.com>